Add base server
The base server consists of a server listening on a user set port
(environment variable ZSSN_PORT).

The base server routes are:
* `POST /survivors` to register a new survivor;
* `GET /survivors` to retrieve a list of all survivors with their IDs
and names;
* `GET /survivors/:id` to retrieve data of a single survivor.

The `api` folder contains the `routes`, `models` and `controllers`
folders for the API management.

The `samples` folder contains a sample json file for registering a new
survivor.