Add reports route
By sending a GET request to the `/reports` route, a basic report of the
human state is returned.

The report contains the total number of survivors, the percentage of
infected and healthy survivors, the average amount of resources per
survivor and the sum of lost points (items) caused by infection.